Add vim, markdown, shell, and whitespace File Type configurations
Modified the bootstrap script to link files based on the structure under
the topic directory.

The File Type plugins belong to both vim and the file type topic. Since
the files are vimfiles (and it's easier to link), I'm putting them in
the vim topic. They _could_ be put under the file type topic if that
becomes easier to manage.